transformation du routeur en module
