+++ index.html
@@ -2,7 +2,8 @@
     <head>
         <!--<link href="https://fonts.googleapis.com/css?family=Nunito:400,600,700" rel="stylesheet">-->
         <!--<link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.6.3/css/all.css" integrity="sha384-UHRtZLI+pbxtHCWp1t77Bi1L4ZtiqrqD80Kn4Z8NTSRyMA2Fd33n5dQ8lWUE00s/" crossorigin="anonymous">-->
-      
+        <script type="module" src="app.js"></script>
+        <script type="module" src="router.js"></script>
         <link rel="stylesheet" type="text/css" href="app.css">
         <link rel="icon" type="image/x-icon" href="img/image.ico" />
     </head>
@@ -10,6 +11,6 @@
         <div id="root"></div>
 
 
-        <script type="text/javascript" src="app.js"></script>
+        <!-- <script type="text/javascript" src="app.js"></script> -->
     </body>
 </html>
